# Conflicts:
#	frontend/package-lock.json
#	frontend/src/app/app.module.ts
#	frontend/src/app/exchangerate/exchangerate.component.css
#	frontend/src/app/exchangerate/exchangerate.component.html
#	frontend/src/app/exchangerate/exchangerate.component.ts
#	server/src/main/java/pl/mjbladaj/zaaw_java/server/dto/CurrencyRate.java
#	server/src/main/java/pl/mjbladaj/zaaw_java/server/rest/SelectedCurrencyRestController.java
#	server/src/main/java/pl/mjbladaj/zaaw_java/server/service/RateService.java
#	server/src/main/java/pl/mjbladaj/zaaw_java/server/service/impl/RateServiceImpl.java
#	server/src/test/java/pl/mjbladaj/zaaw_java/server/service/impl/FreeCurrenciesComRateServiceImplTest.java

+++ server/src/main/java/pl/mjbladaj/zaaw_java/server/converters/RateConverter.java
@@ -10,7 +10,7 @@ public abstract class RateConverter {
          return
                  CurrencyRate
                          .builder()
-                         .rate(rate.getRate())
+                         .rate(rate.getRate().doubleValue())
                          .build();
     }
 }
